[TD-11994]<fix>: taosdump compile on aarch32.

+++ src/taosdump.c
@@ -2336,7 +2336,7 @@ static int64_t writeResultToAvro(
                             __func__, __LINE__, n64tmp);
                     avro_value_append(&value, &secondhalf, NULL);
                     avro_value_set_long(&secondhalf, LONG_MAX);
-                    debugPrint("second half is: %"PRId64"\n", LONG_MAX);
+                    debugPrint("second half is: %"PRId64"\n", (int64_t) LONG_MAX);
 
                     break;
